bldr-build uses a temporary directory when building.
This fixes issues where the src directory is mounted into the container
(and possibly even the Docker host) which leads to umask/permission
inconsistencies. The strategy here is to create a temporary directory,
copy the contents of the directory containing the `Bldrfile` into that
directory and run the build as before.

**Note:** This change updates the UI of `bldr-build` which now accepts
the path to the directory containing the `Bldrfile`, not the `Bldrfile`
itself. The side effect (benefit) is that you no longer have to be in
the directory containing the `Bldrfile`. Also, the current working
directory (`cwd`) is assumed as the default, so less typing!
For example:

    bldr-build /path/to/nginx
    bldr-build ./zlib
    cd ./redis; bldr-build

References: BLDR-17