bldr-build uses a temporary directory when building the Docker image.
This builds on the previous change by only sandboxing the Docker image
creation. This enables package definitions to use relative paths in the
callback hooks without fear that their context will be lost at execution
time. There's also a pleasant symmetry between `build_wrapper` and
`install_wrapper` in the new `dockerfile_wrapper` function which helps
the author of a `Bldrfile` be in the right context at the right moment
in time.

References: BLDR-17