Ensure we have only one signal handler, not many
We accidentally had the creation of the thread that catches signals
inside the main topology event loop. The result was that we would spawn
another thread on every iteration of the loop, and that thread would
dilligently begin listening to signals.

You could see this issue pop up by having spurious "panic" messages
printing to stdout when a signal is received - that's all the other
signal threads panic-ing.