Resolves race condition with waitpid(0)
Bldr has multiple threads going at once. The main thread is running an
event loop that includes a call to `waitpid(0)`, which means wait for
every any child that exits. In a seperate thread, most commonly the
`configuration` thread that watches for filesystem change events for
service reconfiguration, we then execute commands and call rusts
`output()` function. This function calls `waitpid(PID)`, which then
races with the main thread - if the main thread reaps the child before
`output()` can reach it, the thread rust spawns to read the output
panics.

Debugging this required dropping in to `rust-gdb` and doing fun thread
gymnastics. As a side-effect, I've moved all of our calls to
`thread::spawn` to use the `thread::Builder` interface, and set a custom
name per-thread - so you at least have a chance to understand precisely
what thread was the issue.

If anyone needs to do this in the future, you can debug thread panics
with:

```bash
  $ gdb
  > rbreak rust_panic
  # stuff happens
  > info threads
```

Will get you the goods. Good luck, thread hunters!

Love,
Adam