Explicitly name bldr_etcd_1 container name in docker-compose.yml.
As a developer, if I clone this project to directory who's name is *not*
named `bldr`, then docker-compose will choose a different name for the
etcd container. This is currently a problem as the functional tests
required an explicit `--link bldr_etcd_1` when running certain topology
tests. Rather than revist this too deeply at the moment, settting a
clear, consistent name for etcd seems like a reasonable approach.