Communicate over unix domain socket for "Docker in Docker" contexts.
The primary trick here is to bind-mount the default Docker socket
(`/var/run/docker.sock`) into the container that needs to run other
`docker` commands. Because the `-v` flag is concerned with paths local
to the system running the Docker engine and not your workstation this
should be more than safe and frees the project to not be as tightly
coupled to the implemenation of docker-machine.