Virtualbox must die
Sadly, Virtualbox and Docker have some pretty nasty bugs, in particular
with sharing of data from the home directory through to a container. You
can literally have one process get the data that was present when the
container was booted, while another process sees the data that is
current. In the same shell.

[It makes you want to crawl into a hole and hide - the bug is 5 years
old](https://www.virtualbox.org/ticket/9069) - we can safely assume they
are not going to fix it.

![WHY](http://gif.co/trcp.gif)