Add a more useful shell prompt in the chroot & script catpture sessions.
The prompt for the chroot session (for the moment lovingly called the
plan-shell) looks like this:

```
[1][plan-shell:/:0]# cd /plans/runit
[2][plan-shell:/plans/runit:0]# false
[3][plan-shell:/plans/runit:1]# echo "hello"
hello
[4][plan-shell:/plans/runit:0]#
```

It is generally of the form:

```
[N][plan-shell:PWD:X]#
```

where `N` is the command number of the prompt, `PWD` is the current
working directory, and `X` is the exit code of the previous command.

Similarly, the prompt used in a script session (invoked with the
`capture.sh` command) looks like this:

```
[2][capture:glibc:/plans:0]#
```

Where the only difference is the name of the capture session ("glibc" in
the above example).

Note that ANSI prompt coloring is used, but only if the `$TERM` supports
it.