Avoid shell for loops in plans/Makefile.
It turns out that using a shell-style for loop in a Makefile does not
give you the same guarentees as running that same for loop in a shell
script with `set -e` set. The for loop is considered a single line
statement to `make(1)` so any failures inside the loop will not trigger
an abort. This change uses Makefile targets to accomplish the same end
goal.