[plans] Update patchelf to be built with & without bootstrap tools.
The patchelf tools needs libgcc and libstdc++ to run, but we want to use
patchelf in preparing the chef/gcc-libs Plan. Dang, right?

This update modifies the behavior of the chef/patchelf Plan to work
slightly differently if the bootstrap toolchain tools are uses versus in
the (future) default way. Namely: by default we add `chef/gcc-libs` as a
dependency, and if it isn't built yet, we fallback to pulling libgcc and
libstdc++ off the `/usr/lib` path (provided by the `/tools` bootstrap
tools and minimal symlinks).

Trust me, this will either make sense in future changes, or it won't
even matter to future yous. Trust....me, right?