[bldr-build] Add internal `_fix_libtool()` phase to purge sys paths.
A frequently repeating pattern has emerged with software that ships with
Libtool scripts (i.e. `ltmain.sh`): by default they search system paths
for common libraries under `/usr/lib`, `/lib`, etc. As we are striving
for total isolation from any host environment, it is beneficial to
eliminate this source of "system leak".

A default, internal phase to fix this whenever a Libtool script is found
seems reasonable and additionally eliminates more boilerplate in Plan
files.

Thanks to the NixOS/nixpkgs project for the rationale, explanation, and
basis for implementation!

See:
* https://github.com/NixOS/nixpkgs/blob/1a4ab04e298077185dd32d4594825b6e92f60a4e/pkgs/stdenv/generic/setup.sh#L615-L617
* https://github.com/NixOS/nixpkgs/blob/1a4ab04e298077185dd32d4594825b6e92f60a4e/pkgs/stdenv/generic/setup.sh#L628-L631