[acceptance] docker machine setup in default
This commit introduces a new recipe, _docker_machine, that can be used
to set up the node it runs on with all the keys and configuration
required to connect to the docker-machine for the acceptance
environment.

We've lifted this from the provision recipe, and made it a normal recipe
with the chef resources rather than a giant opaque ruby_block.

We also want to assume that the functional recipe should run against the
docker machine we've configured.