[bldr-build] Set `PATH` directly after resolving dependencies.
This change changes the running order of operations in `bldr-build`
slightly to allow for more powerful "Plan upgrades" when setting up a
Plan execution to run with all of the right tools and libraries.

Prior to this change, the `PATH` was adjusted in the
`_build_environment()` phase which is after `do_download()`,
`do_verify()`, and `do_unpack()` have completed. If a Plan requires a
specific version of `wget` or perhaps a tool that can extract 7zip
archives, there is no way to install these these packages before they
are needed. This change sets the `PATH` after the `do_begin()` phase
(which is another oportunity to augment the `$pkg_deps` and
`$pkg_build_deps` arrays), and before `do_download()`.

Note that this affected the `_find_system_commands()` phase in that we
require a GNU version of the `sort` command for the `_latest_package()`
to function correctly. Consequently, an additional function,
`_find_initial_system_commands()`, was added to ensure that any initial
commands were present **before** (and in service of) the package
resolving phase.