[acceptance] Sharing secret through environment
We need to inject the secret - an SSH key - into the Docker container
so that we can clone the private bldr repository. We don't have it
available with the volumes directive in docker-compose.yml because we
can't do a shared folder over to the EC2 node.

This is accomplished by injecting the SSH into the environment that
make executes in during the functional phase. We then check for the
environment variable in the make target, and execute appropriately -
if we have a GITHUB_DEPLOY_KEY we hope for the best that we have the
DELIVERY_GIT_SHASUM and can use the GIT_SSH too.

We also needed an SSH wrapper script, which gets copied over to the
target docker container, so that we can use this deploy key, and not
get prompted for the GitHub SSH host key.