[acceptance] Better logic for bldr src checkout
We want to check that if /src/plans doesn't exist, then cd to the root
directory, clone the bldr repo to /src. If that succeeds, then check
out the shasum for the delivery change, and if re on that commit, make
all the plans. But do that in a subshell if we can cd into the plans
directory.