[bldr-build] Add build time in summary.
Not a large feature, but certainly a useful one--knowing how long a
build took in minutes and seconds. This turns out to be nightmarishly
hard to do in straight Shell/Bash without resorting to the `bc` command
for floating point math. Instead the simplest thing possible was used:
the `$SECONDS` environment variable present in every Bash shell, which
is the number of whole seconds since its process started. Combine this
with a bit of `awk` and we get something reminiscent of the "real" part
of the `time(1)` command. Not that this works even on build failure as
we are hooking into the `_on_exit()` trap function hook. Sweet, no?