Fix a missing Ack timeout
Lets call this the 'Dave is prescient' bug. We were transitioning
packets into the AwaitingAck state, but not timing that state out. The
side effect was that you could easily wind up blocking on an Ack that
never comes, and yet you never mark the other member as suspect. This
commit fixes that.