Isolated members find a way to ressurect
It's possible under many partitioning scenarios that a given node can
find itself completely isolated - it detects all the other members as
down, and it cannot reach another permanent member.

Since an entirely new supervisor always generates an entirely new ID, we
can tell the difference between a supervisor who was gone and came back,
and a supervisor that is new to the ring.

Therefore, if we get completely isolated, we want to get a little
aggressive about trying to find a peer who we can communicate our
alive-ness with. This commit makes it so that if you find yourself
completely isolated, you start sending pings again to all your Confirmed
dead peers.

After the first member responds, you stop treating confirmed dead
members as dead - assuming the rest of the gossip protocol will work out
for you who is actually alive, and who is actually dead.

Live Long and Prosper \\//