Work around a rust perf regression
The rust compiler has a performance regression where it is no longer
re-using it's cache for complex type boundaries. The side effect is our
release builds can take up to 20+ minutes to complete.

While the same regression exists in debug mode, it's far less damaging.
So this PR has us shipping debug builds for now. This wouldn't be a
choice if we were in production anyplace - it would cause a universal
performance regression for every bldr customer. Since we aren't, this is
the easiest way out, while we wait for rust 1.7.

Love,
Adam