[bldr-build] Change latest package logic to look for `MANIFEST`.
Prior to this change, a partially build (but ultimately failed) package
on disk would be returned from `_latest_package()` even if there was no
package at that path. This change now looks for a `MANIFEST` metadata
file in the root of the package path which is then considered a "hit".
While still not 100% correct, there are far less false positive matches
as a result.

Additionally there are 2 refactorings of this function:

* It has been renamed from `_latest_package()` to
  `_latest_install_package()` as a hint to the reader that we are not
  talking about any remote packages--only locally installed ones.
* The case logic complexity has been reduced and we are left with one
  `find | sort | head` pipeline command that all cases flow through.
  This should help to avoid subtle future bugs in one of the case
  branches and hopefully ends up being easier to read.

As this is a published internal function, this is not a breaking change
;)