Development environment with studio
This commit brings the development environment up to date in studio.

To build a package:

```bash
$ make pkg-shell
$ studio enter
$ make gpg
$ build plans/redis
```

To upload the resulting package

```bash
$ studio enter
$ ./plans/support/cheap-upload.sh PKG
```

To create a docker container of a package, either local or remote:

```bash
$ studio enter
$ dockerize chef/redis
```

To develop bldr itself, just work like you always did. If you want to,
for example, test that redis is working with your development version of
the supervisor:

```bash
$ bpm install chef/redis
$ ./target/debug/bldr start chef/redis
```

Will work just fine (as will running bldr on other host operating
systems, cause thats all we're up to).

This commit does not fix the test suite - we will still be failing
functionals. Next PR. :)