Reduce the complexity of the root Makefile.
This change puts our root Makefile, Dockerfile, and docker-compose.yml
on a diet. Only what is purely required is specified and any older
targets are purged.

The biggest change is that the default `make` target is not `make build`
which creates the Docker development image and compiles the supervisor
Rust project.

Additionally, `make doc-serve` has been renamed to `make docs-serve`.