bldr install now honors installed packages
When running `bldr install chef/redis`, we previously would download all
dependencies regardless of whether they were installed. This changes
that behavior, so that we check to see if the dependency we need is
already installed, and if it is, we move on.

This also updates some behavior in `Package::load`, where if you passed
a fully qualified `PackageIdent`, we would always return a `Package`
struct, even if that package is not installed. It now returns an error
in that case.