[delivery] Only fully rebuild devshell image once a day.
This change makes our "devshell" Docker image rebuilding much less
aggressive when running in Delivery. Prior to this change, the devshell
Docker image was removed completely and then rebuilt from scratch for
every `unit`, `lint`, and `functional` execution. While this made the
image as fresh as possible, there is a fair amount of wasted and waiting
time on every execution.

With this change, the "devshell" image is fully rebuilt if the existing
one is older than 24 hours. Otherwise, `refresh=true` is passed to the
Makefile targets which runs the phase with through a `docker build`
first, relying on previous image cache layers. This means that if a
change involves updates to the Dockerfile, then the "devshell" image
will be refreshed on the first execution.