Audit all Error types in `bldr` crate.
This change compares all error types in the `bldr` crate against what is
currently in the `bldr_core` crate and drops any error types that are no
longer used due to refactoring, reimplementation, etc. Additonally, the
Error enums are alphabetized in order to make future comparison audits a
simpler task. Also, this means there is one natural place for a net-new
enum: where it sits lexically.