[Makefile] Re-order build order of components and trim `make all`.
This change runs the test suites in an order from lower level components
up to higher level components. Additionally, non-runnable compoenents
(i.e. library-only crates) aren't directly built with the `make all`
target. This is because these library compoenents are built when
building the `hab-sup` and `hap-depot` binaries. This cuts the run time
of `make all` roughly in half--on my system from ~10 minutes to ~5
minutes.