[delivery] Update `functional` phase to remove the `tee` pipe.
As we are no longer compiling a lot of software a supremely long run
time of our functional suite should be a failure and should normally
output as it goes. Worse, the error of the `make` target may be getting
ignored in favor of `tee`'s exit code which results in an ever-green
`functional` phase.