[hab-sup] Allow the Supervisor to run in a minimal environment.
This change adds the `chef/busybox-static` dependency to `chef/hab-sup`
to use its `chpst` command on `PATH` to replace the one formerly
provided in the `chef/runit` dependency.

Additionally this change sets up the transitionaly Habitat package
`chef/hab-sup` as the canoncial Supervisor package which is used to
determine the runtime path when issuing `hab-sup start` commands.