Origin Creation
When you are signed in, do not have an origin, and click the Add Origin
link or go to a page that requires you to be signed in, show the form to
create an origin.

Right now the only valid name is "smith" and it doesn't actually create
the origin on the server or persist it after page loads.

There is a bug that will redirect you to the sign in page when you're
already signed in, but that's another ticket.

Also:

* Fix the height of the body container
* Have a signing in indication when sign in is happening