[hab] Create an additional statically-built Plan which uses musl.
This change introduces a new Plan (under `components/hab/static`) which
uses a musl-based libc chain of dependencies and produces a ~6.0mb
statically built binary which should be portable enough to work in our
stage1 Studios and potentially on third-party 64-bit Linux operating
systems.

As a hedge on not getting everything totally right the first time, the
default build is an improved glibc-based, dynamic binary as before,
although using the `cargo build --release` flag for a tighter resulting
binary.