[plan-build] Fix checksum generation in MANIFEST file.
The previous implementation was passing a null byte delimited list to
sort which isn't valid and also causes the pipe (and the subshell) to
fail. This appears to have triggered the `ERR` trap condition in the
main program and write logging information to the stdout of the
subshell. For example:

```
0ddd8455b2ec1d98844d8a051939acaa5cb2b9a590bcfacedea86ec8454537c6  /hab/pkgs/core/glibc/2.22/20160415170530/lib/libnss_compat-2.22.so
9855b2fd764bcb0bad6cf07f9803cb29cb5c73d020c22fa5ae3c7e67a0e7da73  /hab/pkgs/core/glibc/2.22/20160415170530/lib/libnsl-2.22.so
5b5aff5fbb6be7dc5cdd51ac41b6039f9109908d03962ea94a2337a78b8b06c8  /hab/pkgs/core/glibc/2.22/20160415170530/lib/librt.a
181164699aa1b230962ed82a97fa5b0b705ee2eea21d13baaaeffcec08a60335  /hab/pkgs/core/glibc/2.22/20160415170530/DEPS
   glibc: Build time: 10m47s
   glibc: Exiting on error
```

Worse, the logging lines are ansi-colored, meaning that the output
contains non-stric ASCII characters or at a minimum multi-byte
characters such as the ansi escape code `U+001B` (src:
http://www.fileformat.info/info/unicode/char/1b/index.htm)

This fix drops the `find -print0/xargs -0` pairing and instead pipes the
sort output to a stream-consuming `while read line` style to get around
the original command-line-too-long issue.