[depot] Add more rigid failure detection on upload to return HTTP/500.
This change attempts to get the Depot server to explicitly fail uploads
when it knows that it can't complete the operation. Prior to this change
we were crashing the request thread (via `unreachable!()`), and in this
change we are returning the reponse as an Internal Server Error
(HTTP/500). This allows us to write client logic that will crash out if
the server does.