Fix redirection on page load bug
Make it so when you load the page, instead of checking for isSignedIn
(which will always be false on initial page load, since we're not
finished authenticating yet at that point), check for the presence of a
token in session storage.

Also simplify some of the token handling and put a sign out button on
the sign in page when signed in instead of insisting on redirecting away
from it.