[sup] Update how Supervisor sets `PATH` on start.
This change uses a more robust strategy for setting the `PATH`
environment variable before a supervised process is start via the
`start` subcommand.

Prior to this change, an install of the supervisor package was required
in order for the Supervisor to start any service, which included
running the Supervisor in the source tree for purposes of development.
This change makes it more explict what we are after in the supervisor's
`PATH` entry: a minmal BusyBox userland.

There is a series of fallback strategies used here in order to find a
usable BusyBox installation. The general strategy is the following:

* Are we (the Supervisor) running inside a package?
    * Yes: use the BusyBox release describes in our `DEPS` metafile
           & return its `PATH` entries
    * No
        * Can we find any installed BusyBox pacakge?
            * Yes: use the latest installed BusyBox release
              & return its `PATH` entries
            * No
                * Is the `busybox` binary present on `$PATH`?
                    * Yes: return the parent directory which holds
                           the `busybox` binary
                    * No: out of ideas, so return an error after warning
                          the user we're done

The end result is that an installation of the Supervisor package is no
longer required in a development environment--only an installation of
the `core/busybox-static` package. Additionally, the precise release of
the `core/busybox-static` dependency will be used when running the
Supervisor out of a pacakge (its natural home).