[core] Add `env::var` & `env::var_os` functions to check for empty vars.
The standard library will consider an empty-but-set environment
variable a legit result. After inspecting the codebase, it turns out
that is not the expectation in most of our code. Indeed, a set-but-empty
`HAB_DEPOT_URL` is not valid and should be ignored.

This change introduces 2 functions which shadow their standard library
cousins providing this additional check.