Add support to encrypt messages between Supervisors with a shared symmetric key.
Wire level message communication
--------------------------------

This feature adds a new concept, called a Ring (or Ring key) to a
Supervisor when started. A Ring refers to a set of Supervisors that
communicate together in order to coordinate services, configuration,
etc. A Ring key is a shared symmetric key that all Supervisors in a Ring
must posess, and with which they will use to 1) encrypt all gossip
messages on send and 2) decrypt all gossip messages on receive. A simple
`WireMessage` struct has been introduced to encapsulate the behavior of
plain vs. encrypted wire mesages.

An encrypted wire message has several fields that are send in the clear
(i.e. not encrypted) to compensate for the lack of protocol negotiation
which exists in other systems such as StartTLS, OpenSSH, etc. These
fields are:

* `format` which denotes if the message is plaintext or encrypted
* `version` which is a protocol message string, designed to be lexically
  sortable as a hedge against future breaking changes
* `key` which is the logical name of the Ring key used to encrypt the
  message, including the revision number
* `nonce` a cryptographic
  [nonce](https://en.wikipedia.org/wiki/Cryptographic_nonce) which is
  associated with the ciphertext
* `msg_bytes` which is the ciphertext, that is, the encrypted message

Note that for a plaintext message the same fields are transmitted but
are left purposefully blank:

* `format` which is set to `Plain`
* `version` which is the same as above
* `key` which is empty or unset
* `nonce` which is empty or unset
* `msg_bytes` which contains the raw contents of the plaintext message

In the case of encrypted messages, only the specific key with revision
can be used to decrypt the ciphertext. Additionally, the same nonce
value is required to decrypt the message as each encrypted message will
have its own uniquely generated random nonce value.

Hab CLI ring key creation
-------------------------

The `hab` command line tool has a subcommand which will generate a ring
key, used to encrypt communication of Supervisor. The ring key must have
a name, which is a logical representation of the Ring. This could be as
refer to an environment such as `"staging"` or `"production"`, a
geographic region such as `"us-west"` or `"sea"`, or as simple as
`"default"`--the value is arbitrary.

Basic usage:

    hab ring key generate titan

This will create a secret key file in your local key cache, which by
default is in `/hab/cache/keys`.

Supervisor user experience
--------------------------

The Supervisor's (the `hab-sup` binary) `start` subcommand has been
augmented to include a `--ring` option. The value is the key name for a
ring key which exists in the Supervisor's key cache. That is, a
corresponding ring key must exist (by default) in the Supervisor's local
`/hab/cache/keys` directory. If a suitable corresponding key cannot be
found on startup, the Supervisor will abort and exit with an error
message. This is to prevent a Supervisor from starting in an
unrepairable state.

Basic usage:

    hab-sup start --ring titan core/redis

Joining an existing ring peer:

    hab-sup start --ring titan --peer 172.17.0.3 core/redis

Applying runtime configuration using Hab CLI
--------------------------------------------

The same pattern applies to the `hab` binary as it applies to the
Supervisor. A new option `--ring` has been added that allow `hab` to
encrypt its message into the Ring. As with the Supervisor the precise
ring key name with revision is required for the CLI to talk with the
Ring.

Basic usage:

    echo 'loglevel = "debug"' | hab config apply --ring titan --peer 172.17.0.3 redis.default 1

Uploading file using Hab CLI
----------------------------

Uploading files to the Ring via the `upload file` subcommand has also
been updated with a `--ring` option to allow `hab` to encypt the file
upload message into the Ring.

Basic usage:

    hab file upload --ring possums --org acme --peer 172.17.0.3 redis.default /tmp/laika.txt 2 jcash

In the example above:

* `--org acme` refers to a runtime org called acme
* `redis.default` refers to the target service group for the file
* `/tmp/laika.txt` refers to the file itself, available on local disk to
  `hab`
* `jcash` refers to the user which will sign-and-encrypt this file. Note
  that file encryption is a seperate concern, orthogonal to wire or
  message encryption.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Pull request: #436
Approved by: fnichol