Update Habitat Artifact (.hart) header information.
Note that this will invalidate all previously build artifacts.

This change adds a simple lexically sortable version string on the first
plaintext line of the artifact. It additionally adds a blank line
bewteen the header/prelude and the data binary stream. In other words
external tooling could opt to check for the first blank line or 2
newline characters as the delimiter.