[hab,sup,depot] Use the version/release info from the Plan if possible.
This change effectively uses the `$pkg_version` and `$pkg_release`
values from `hab-plan-build` when writing out the version str constant
in the resulting binaries. It does this via a very simple `build.rs`
program in each component which looks for a `$PLAN_VERSION` environment
variable. If this variable does not exist, it falls back to the default
`$CARGO_PKG_VERSION` which pulls directly from the component's Cargo
manifest.

The end result is that components built into a pacakge will report their
version/release information which is exactly identical to the
information in the package identifier. Additionally, this should help
track the relative age of any future static binaries that may get
deployed into the wild (at the moment, only `hab` is being statically
built).

For example:

```
> /hab/pkgs/core/hab-static/0.4.0/20160423180136/bin/hab --version
hab 0.4.0/20160423180136

```