[bpm] Update `hab-bpm install` to understand updated artifact format.
Due the changes in #425, the header portion of a Habitat Artifact has
grown from the first 3 lines to 5 lines, the last of which is an empty
line. In order for the "cheap" (i.e. dirty, evil, not recommended)
install method that `hab-bpm` currently uses to work, we need to start
reading the tar archive starting on line 6 now, not 4.

Hm, the commit message is far longer than the fix. Funny how that works.