Update Studio install strategy & trim down Dockerfile.
This change alters the Studio's install script to use more Habitat. It
now consumes a Habitat artifact file for `core/hab-bpm` and extracts
under the `HAB_PKG_ROOT_PATH` which means that it is upgrade-ready.

For the moment, the `core/hab-studio` and `core/busybox-static` packages
are installed directly from an S3 bucket which will change once a
default Depot service has been relaunched.