[core] Re-enable artifact reading test with new artifact format.
This change adds a very small plan to produce a very small artifact
which was added into core's fixtures directory along with the origin
public and secret signing keys. This should allow future maintainers to
update this fixture if and when any formatting updates take place. There
is also now a concrete example of an origin keypair available for other
tests.