[plan-build] Allow `attach` session to handle non-zero exit cmds.
This fixes a long outstanding issue where the debugging session would
terminate the first time a user entered a command that returned
non-zero. This would include typos, etc. and doesn't make the attach
debugger feel very friendly. The ended up being clearing the on-exit
traps and allowing for non-zero command returns by setting `set +e` for
the duration of the session.